backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
python-telegram-bot==20.8
anthropic==0.18.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
alembic==1.13.1
python-multipart==0.0.6
aiosqlite==0.19.0
httpx==0.26.0
